Create DocsToSheet.py
This small script should read all the docs in GDOC format, in your drive and transfer the content to a spreadsheet.

Don't share access to this file if used in Colab, since it requires authentication and you would be exposing your drive to whoever you shared it with. 

I used it to parse my data I downloaded from Facebook.
Use at your own responsibility. 

Attention, this is terrible and has no error handling. 
Peace.